+++ apps/web/src/app/[lang]/app/[type]/operations/details/[tagId]/_components/totals.tsx
